pyproject.toml
[tool.poetry]
name = "isaric2omop"
version = "0.1.0"
description = ""
authors = ["anastasianayden <anastasia.nayden@health-ri.nl>"]
readme = "README.md"
packages = [{include = "isaric2omop"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.2"
sqlalchemy = "^1.4.44"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
